+++ app/en/(auth)/success/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Header } from "@/components/spanish/layout/header";
+import { Header } from "@/components/en/layout/header";
 import { supabase } from "@/lib/supabase";
 
 export default function SuccessPage() {
@@ -16,16 +16,21 @@ export default function SuccessPage() {
   const [definitions, setDefinitions] = useState<{ [key: string]: { definition: string; partOfSpeech: string } }>({});
   const [imageUrl, setImageUrl] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
+  const [story, setStory] = useState<string | null>(null);
+  const [audioSrc, setAudioSrc] = useState<string | null>(null);
 
   // Fetch words from Supabase database
   useEffect(() => {
+    localStorage.setItem("language", "en");
+    
     const fetchWords = async () => {
       const { data, error: userError } = await supabase.auth.getUser();
 
       const userId = data.user?.id;
       const sharedUserId = process.env.NEXT_PUBLIC_SHARED_USER_ID;
+      const selectedLanguage = localStorage.getItem("language") || "en"; 
       
-      const { data: wordsData, error } = await supabase.from("messages").select("text").in("uid", [userId, sharedUserId]);;
+      const { data: wordsData, error } = await supabase.from("messages").select("text").in("uid", [userId, sharedUserId]).eq("language", selectedLanguage);
       if (error) {
         console.error("Error fetching words:", error);
       } else {
@@ -56,8 +61,11 @@ export default function SuccessPage() {
       console.error("Error adding new word:", userError);
       return;
     }
+
+    const selectedLanguage = localStorage.getItem("language") || "en";
+
     const { error } = await supabase.from("messages").insert([
-      { text: newWord, uid: user.id }
+      { text: newWord, uid: user.id, language: selectedLanguage }
     ]);
     if (error) {
       console.error("Error adding new word:", error);
@@ -71,9 +79,11 @@ export default function SuccessPage() {
   const handleAddHoveredWord = async () => {
     if (!hoveredWord?.word) return;
 
+    const selectedLanguage = localStorage.getItem("language") || "en";
+
     // Add the word to the database if it's not already present
     if (!words.includes(hoveredWord.word)) {
-      const { error } = await supabase.from("messages").insert([{ text: hoveredWord.word }]);
+      const { error } = await supabase.from("messages").insert([{ text: hoveredWord.word, language: selectedLanguage }]);
       if (error) {
         console.error("Error adding hovered word:", error);
       } else {
@@ -107,6 +117,7 @@ export default function SuccessPage() {
     const data = await response.json();
     if (data?.story) {
       setGeneratedStory(data.story);
+      setStory(data.story);
       applyHighlighting(data.story);
       await generateImageFromStory(data.story);
     } else {
@@ -124,10 +135,14 @@ export default function SuccessPage() {
         body: JSON.stringify({ story }),
       });
 
-      const data = await response.json();
-      if (data?.imageUrl) {
-        setImageUrl(data.imageUrl);
-      }
+      // Convert response to Blob
+      const blob = await response.blob();
+
+      // Convert Blob to Object URL
+      const imageObjectURL = URL.createObjectURL(blob);
+      console.log("Generated Image URL:", imageObjectURL);
+  
+      setImageUrl(imageObjectURL);
     } catch (error) {
       console.error("Error generating image:", error);
     } finally {
@@ -168,6 +183,40 @@ export default function SuccessPage() {
 
     fetchDefinition();
   }, [hoveredWord]);
+
+
+  // this function converts the text to speech
+  const handleConvertToSpeech = async () => {
+
+      if (!story) return alert("Generate a story first!");
+
+      try {
+        const response = await fetch("/api/generate-full-audio", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ text: story }),
+        });
+    
+        if (!response.ok) {
+          console.error("Error fetching audio:", response.statusText);
+          return alert("Failed to generate audio.");
+        }
+    
+        const audioBlob = await response.blob();
+        const audioUrl = URL.createObjectURL(audioBlob);
+        setAudioSrc(audioUrl); 
+
+        console.log("Audio generated:", audioUrl);
+      }
+      
+      catch (error) {
+        console.error("Error in TTS request:", error);
+      }
+    
+
+  };
+
+  
 
   return (
     <div className="min-h-screen flex flex-col bg-background">
@@ -219,9 +268,20 @@ export default function SuccessPage() {
               <Button variant="outline" className="mb-4 border-purple-500" onClick={handleGenerateStory}>
                 + Generate Story
               </Button>
-              <Button variant="secondary" className="bg-purple-500 text-white hover:bg-purple-600">
-                Audio
-              </Button>
+
+              {generatedStory && (
+                <Button variant="secondary" className="bg-purple-500 text-white hover:bg-purple-600" onClick={handleConvertToSpeech}>
+                  Read Aloud
+                </Button>
+              )}
+
+              
+              {audioSrc && (
+                <audio key={audioSrc} controls autoPlay className="mt-0">
+                <source src={audioSrc} type="audio/mpeg" />
+                  Your browser does not support the audio element.
+              </audio>
+            )}
             </div>
 
             {/* Story Output */}
