+++ app/es/(auth)/success/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Header } from "@/components/layout/header";
+import { Header } from "@/components/spanish/layout/header";
 import { supabase } from "@/lib/supabase";
 
 export default function SuccessPage() {
@@ -16,18 +16,19 @@ export default function SuccessPage() {
   const [definitions, setDefinitions] = useState<{ [key: string]: { definition: string; partOfSpeech: string } }>({});
   const [imageUrl, setImageUrl] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
-  const [story, setStory] = useState<string | null>(null);
-  const [audioSrc, setAudioSrc] = useState<string | null>(null);
 
   // Fetch words from Supabase database
   useEffect(() => {
+    localStorage.setItem("language", "es");
+    
     const fetchWords = async () => {
       const { data, error: userError } = await supabase.auth.getUser();
 
       const userId = data.user?.id;
       const sharedUserId = process.env.NEXT_PUBLIC_SHARED_USER_ID;
-      
-      const { data: wordsData, error } = await supabase.from("messages").select("text").in("uid", [userId, sharedUserId]);;
+      const selectedLanguage = localStorage.getItem("language") || "es"; 
+
+      const { data: wordsData, error } = await supabase.from("messages").select("text").in("uid", [userId, sharedUserId]).eq("language", selectedLanguage);
       if (error) {
         console.error("Error fetching words:", error);
       } else {
@@ -58,8 +59,11 @@ export default function SuccessPage() {
       console.error("Error adding new word:", userError);
       return;
     }
+
+    const selectedLanguage = localStorage.getItem("language") || "es";
+
     const { error } = await supabase.from("messages").insert([
-      { text: newWord, uid: user.id }
+      { text: newWord, uid: user.id, language: selectedLanguage  }
     ]);
     if (error) {
       console.error("Error adding new word:", error);
@@ -73,9 +77,11 @@ export default function SuccessPage() {
   const handleAddHoveredWord = async () => {
     if (!hoveredWord?.word) return;
 
+    const selectedLanguage = localStorage.getItem("language") || "es";
+
     // Add the word to the database if it's not already present
     if (!words.includes(hoveredWord.word)) {
-      const { error } = await supabase.from("messages").insert([{ text: hoveredWord.word }]);
+      const { error } = await supabase.from("messages").insert([{ text: hoveredWord.word, language: selectedLanguage }]);
       if (error) {
         console.error("Error adding hovered word:", error);
       } else {
@@ -109,7 +115,6 @@ export default function SuccessPage() {
     const data = await response.json();
     if (data?.story) {
       setGeneratedStory(data.story);
-      setStory(data.story);
       applyHighlighting(data.story);
       await generateImageFromStory(data.story);
     } else {
@@ -127,14 +132,10 @@ export default function SuccessPage() {
         body: JSON.stringify({ story }),
       });
 
-      // Convert response to Blob
-      const blob = await response.blob();
-
-      // Convert Blob to Object URL
-      const imageObjectURL = URL.createObjectURL(blob);
-      console.log("Generated Image URL:", imageObjectURL);
-  
-      setImageUrl(imageObjectURL);
+      const data = await response.json();
+      if (data?.imageUrl) {
+        setImageUrl(data.imageUrl);
+      }
     } catch (error) {
       console.error("Error generating image:", error);
     } finally {
@@ -175,40 +176,6 @@ export default function SuccessPage() {
 
     fetchDefinition();
   }, [hoveredWord]);
-
-
-  // this function converts the text to speech
-  const handleConvertToSpeech = async () => {
-
-      if (!story) return alert("Generate a story first!");
-
-      try {
-        const response = await fetch("/api/generate-full-audio", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ text: story }),
-        });
-    
-        if (!response.ok) {
-          console.error("Error fetching audio:", response.statusText);
-          return alert("Failed to generate audio.");
-        }
-    
-        const audioBlob = await response.blob();
-        const audioUrl = URL.createObjectURL(audioBlob);
-        setAudioSrc(audioUrl); 
-
-        console.log("Audio generated:", audioUrl);
-      }
-      
-      catch (error) {
-        console.error("Error in TTS request:", error);
-      }
-    
-
-  };
-
-  
 
   return (
     <div className="min-h-screen flex flex-col bg-background">
@@ -260,20 +227,9 @@ export default function SuccessPage() {
               <Button variant="outline" className="mb-4 border-purple-500" onClick={handleGenerateStory}>
                 + Generate Story
               </Button>
-
-              {generatedStory && (
-                <Button variant="secondary" className="bg-purple-500 text-white hover:bg-purple-600" onClick={handleConvertToSpeech}>
-                  Read Aloud
-                </Button>
-              )}
-
-              
-              {audioSrc && (
-                <audio key={audioSrc} controls autoPlay className="mt-0">
-                <source src={audioSrc} type="audio/mpeg" />
-                  Your browser does not support the audio element.
-              </audio>
-            )}
+              <Button variant="secondary" className="bg-purple-500 text-white hover:bg-purple-600">
+                Audio
+              </Button>
             </div>
 
             {/* Story Output */}
